backend/requirements.txt
flask
pdfminer.six
tabula-py
PyMuPDF
pandas
pytesseract
Pillow
